Windows should now work completely.
Add material index support and optimize ray query paths

- Introduced `materialIndex` in `InstanceData` to streamline per-instance material assignments in ray queries.
- Enhanced Vulkan descriptor bindings, adding new storage buffers for material and geometry data (Bindings 12 and 13).
- Extended `ModelLoader` with material index mapping and lookup functions for improved glTF material handling.
- Implemented robust material index resolution logic and authoritative mappings in TLAS build paths.
- Improved handling of alpha-masked and environment objects with dynamic ray-query classifications.
- Optimized geometry info and material buffer generation for better consistency in ray tracing workflows.
- Addressed descriptor flag size mismatches and resolved instances of unsafe type casting in Vulkan resource bindings.

+++ attachments/simple_engine/physics_system.cpp
@@ -1276,7 +1276,7 @@ void PhysicsSystem::SimulatePhysicsOnGPU(const std::chrono::milliseconds deltaTi
 
   // Step 1: Integrate forces and velocities
   vulkanResources.commandBuffer.bindPipeline(vk::PipelineBindPoint::eCompute, *vulkanResources.integratePipeline);
-  vulkanResources.commandBuffer.dispatch((rigidBodies.size() + 63) / 64, 1, 1);
+  vulkanResources.commandBuffer.dispatch(static_cast<uint32_t>((rigidBodies.size() + 63) / 64), 1, 1);
 
   // Memory barrier to ensure integration is complete before collision detection
   vk::MemoryBarrier memoryBarrier;
@@ -1293,7 +1293,7 @@ void PhysicsSystem::SimulatePhysicsOnGPU(const std::chrono::milliseconds deltaTi
 
   // Step 2: Broad-phase collision detection
   vulkanResources.commandBuffer.bindPipeline(vk::PipelineBindPoint::eCompute, *vulkanResources.broadPhasePipeline);
-  uint32_t numPairs = (rigidBodies.size() * (rigidBodies.size() - 1)) / 2;
+  uint32_t numPairs = static_cast<uint32_t>((rigidBodies.size() * (rigidBodies.size() - 1)) / 2);
   // Dispatch number of workgroups matching [numthreads(64,1,1)] in BroadPhaseCS
   // One workgroup has 64 threads, each processes one pair by index
   uint32_t broadPhaseThreads = (numPairs + 63) / 64;
